Set exposures primary key to be longer

+++ winterdrp/pipelines/summer/summer_files/schema/exposures.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS exposures (
-    expid SERIAL PRIMARY KEY,
+    expid BIGINT PRIMARY KEY,
     obsdate INT,
     obsID INT,
     itid INT,
